refactor: split backends/macos.py (1,065 lines) into focused submixins (fixes #862)
